major refactoring of code

+++ csv.cpp
@@ -5,6 +5,13 @@
 
 using namespace std;
 
+string file_input(){
+    string filename;
+    cout << "Enter the file name (without extension): ";
+    cin >> filename;
+    filename.append(".csv");
+    return(filename);
+}
 int main(){
     int cont = 1;
     Remove r;
@@ -17,11 +24,15 @@ int main(){
             case 'r':
             {
                 Read fileObj;
+                // call the file open function from the 
+                // read class with the generic user question
+                fileObj.File_O(file_input());
                 break;
             }
             case 'w':
             {
                 Write newFile;
+                newFile.populate_file(file_input());
                 // keep track of the created files for garbage collection
                 // remove if file is needed
                 files.insert(files.end(), newFile.created_file.begin(),newFile.created_file.end());
@@ -36,8 +47,14 @@ int main(){
         cout << "input another file? (y/n): ";
         cin >> response;
         if(response != 'y'){
-            for(int i = 0; i < files.size(); i++)
-                r.remove_file(files[i]);                      
+            if(files.size() != 0){
+                cout << "delete created files? (y/n): ";
+                cin >> response;
+                if(response != 'n'){
+                    for(int i = 0; i < files.size(); i++)
+                        r.remove_file(files[i]);                      
+                }
+            }
             cont = 0;
         }
     }
